finishing out the upvote feature
So making a list of usernames on a post that liked it and checking that to enforce no-double-upvoting is pretty sketch, I won't lie. I'm not a big database guy so I didn't remember the "right" way to do that off the top of my head and Doom Eternal was about to come out so I just slapped that together and called it a day. I'll make a bug for it so we don't forget to fix that. I _think_ the correct way to do that is to make a join table (or a relationship table? I don't remember the term) where we map users -> upvoted posts, and then we can search that for our userid (NOT USERNAME -- never forget PSN and the original sin of using names for everything!) and check if this post id is in that list.

Actually, typing that solution out, it sounds right. I'll drop a note in the bug.

- made the upvote button only work for logged in users
- added upvote information to the posts page on load
- cleaned up frontend script